+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 5.7.2 (in progress)
 
-* Your contribution here
+##### Bug fixes / Improvements
+* [#1614](https://github.com/oshi/oshi/pull/1614): Simplify Windows version parsing using JDK properties - [@dbwiddis](https://github.com/dbwiddis).
 
 # 5.7.0 (2021-04-01), 5.7.1 (2021-04-15)